% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intialize.R
\name{compute_parameters_initial_values}
\alias{compute_parameters_initial_values}
\title{Given a fastgllvm object, compute good intial values.}
\usage{
compute_parameters_initial_values(fastgllvm, target = NULL, rescale = T)
}
\arguments{
\item{fastgllvm:}{an object of class ''fastgllvm''}

\item{target:}{if non NULL, must be a matrix of loadings: will perform a Procrustes rotation to target}

\item{rescale:}{if true, both Z and A will be rescaled so that Z\%*\%t(A) remains unchanged but Z has variance identity.}
}
\description{
Given a fastgllvm object, compute good intial values.
}
