% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi.R
\name{compute_psi}
\alias{compute_psi}
\title{Computes the psi function}
\usage{
compute_psi(Y, Z, X, A, B, phi, families)
}
\arguments{
\item{Y:}{n*p matrix of responses}

\item{Z:}{n*q matrix of latent variables}

\item{X:}{n*k design matrix}

\item{A:}{p*q loadings matrix}

\item{B:}{p*k coefficients matrix}

\item{family:}{known parameters}
}
\description{
Computes the psi function
}
