% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Z.R
\name{compute_Z}
\alias{compute_Z}
\title{Compute the MAP}
\usage{
compute_Z(
  fg,
  start = NULL,
  maxit = 10,
  thresh = 0.001,
  lambda = 1,
  save = F,
  verbose = F
)
}
\arguments{
\item{start:}{initial estimate of the MAP. If NULL, initial values will be computed depending based on an initial estimate of mu.}

\item{maxit:}{maximum number of Newton iterations}

\item{thresh:}{threshold to stop the routine; see}

\item{lambda:}{the regularization parameter}

\item{save:}{iterations are saved}

\item{verbose:}{if TRUE, documents and prints the progress}
}
\description{
For each observational unit i, computes using the Newton method the maximum a posteriori value, i.e. the values that maximizes the distribution of Z|Y_i for given parameter values.
}
\details{
For computational speed, the convergence is checked for each observational unit \code{i} separately, and computations are vectorized across observations as much as possible.
The \code{lambda} parameter is the regularization parameter. Setting it to 1 (the fault) returns the MAP. Higher values regularizes the values towards 0.
}
