% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi.R
\name{compute_zstar}
\alias{compute_zstar}
\title{Compute zstar}
\usage{
compute_zstar(
  Y,
  A,
  phi,
  XB,
  families,
  dims,
  start = NULL,
  maxit = 100,
  thresh = 0.001,
  save = F
)
}
\arguments{
\item{families:}{a named list of indices corresponding to the families and their respective object}

\item{start:}{starting values for zstar}

\item{simplify_hessian:}{if TRUE, only the diagonal of the hessian is used
(not yet implemented)}

\item{line_search:}{if TRUE, adds a line_search. This makes the convergence
more robust, at the potential cost of computation speed.}

\item{maxit:}{maximum iterations to compute zstar}

\item{thres:}{threshold to stop the routine}
}
\description{
Implemented using Newton method. Starting values are important.
}
