% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Zi.R
\name{compute_zstar_i}
\alias{compute_zstar_i}
\title{Compute zstar}
\usage{
compute_zstar_i(
  Y_i,
  A,
  phi,
  XB_i,
  families,
  start = NULL,
  maxit = 100,
  thresh = 0.001,
  save = F,
  scale = F,
  Miss_i = NULL
)
}
\arguments{
\item{families:}{a named list of indices corresponding to the families and their respective object}

\item{start:}{starting values for zstar}

\item{simplify_hessian:}{if TRUE, only the diagonal of the hessian is used
(not yet implemented)}

\item{line_search:}{if TRUE, adds a line_search. This makes the convergence
more robust, at the potential cost of computation speed.}

\item{maxit:}{maximum iterations to compute zstar}

\item{thres:}{threshold to stop the routine}

\item{Y_i:}{a vector of observations}

\item{A:}{the loadings (act as the covariates)}

\item{XB_i:}{the offset}

\item{Miss_i:}{the missing values}
}
\description{
Implemented using Newton method. Starting values are important.
}
