% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Zi.R
\name{compute_zstar_i_starting_values}
\alias{compute_zstar_i_starting_values}
\title{Compute Starting values for zstar.}
\usage{
compute_zstar_i_starting_values(Y_i, A, XB_i, families)
}
\description{
Good  starting values are used to speed up / robustify the convergence. They are computed
by first transforming the response and then taking the OLS.
}
\details{
NAs are automatically taken care of by lm.
}
