% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi.R
\name{compute_zstar_starting_values}
\alias{compute_zstar_starting_values}
\title{Compute Starting values for zstar.}
\usage{
compute_zstar_starting_values(Y, A, XB, families, dims)
}
\description{
Good  starting values are used to speed up / robustify the convergence. They are computed
by first transforming the response and then taking the OLS.
}
