% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgllvm.R
\name{fastgllvm.fit}
\alias{fastgllvm.fit}
\title{Workhorse function to fit a fastgllvm model}
\usage{
fastgllvm.fit(fg, controls, verbose, hist, parameters.init = NULL)
}
\arguments{
\item{fg:}{an object of type "fastgllvm"}

\item{controls:}{a list of parameters controlling the fitting process}
}
\description{
Workhorse function to fit a fastgllvm model
}
