% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgllvm.R
\name{fastgllvm}
\alias{fastgllvm}
\title{Fit a fastgllvm}
\usage{
fastgllvm(
  Y,
  q = 1,
  family = "gaussian",
  X = NULL,
  intercept = T,
  Z.init = NULL,
  parameters.init = NULL,
  controls = list(),
  verbose = F,
  hist = F
)
}
\arguments{
\item{Y:}{a \code{n} times \code{p} matrix of observations}

\item{q:}{the number of factors}

\item{family:}{specifies the family. See below for more information.}

\item{X:}{either \code{NULL} or a \code{n} timeis \code{k} matrix of covariates.}

\item{intercept:}{a boolean (default:TRUE) indicating whether an intercept should be included in the model. If \code{X} is supplied and an intercept is desired, the first column of \code{X} must be a vector of ones to model the intercept.}
}
\description{

}
\details{
The implemented families currently are "gaussian", "binomial", or "poisson.
The \code{family} argument can be one of the following:
\itemize{
\item a string, the name of one of the implemented families.
\item a vector of strings of size \code{p}, each of which being the name of the implemented families.
\item a list: \code{list(objects, id, vec)}, where \code{objects} is a named list of objects of type "family",
\code{id} is a named list of integer vectors specifying the column numbers corresponding, whose names correspond to those of \code{objects}, and
\code{vec} is a named list of string vectors whose elements are the names corresponding to the \code{objects}$family name. This can be useful to specify
different link functions for the families.
}
}
