% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{fastgllvm_oop}
\alias{fastgllvm_oop}
\title{Returns a fitted object of class "fastgllvm".}
\usage{
fastgllvm_oop(
  Y,
  q = 1,
  X = 0,
  family = binomial(),
  method = "SA",
  H = 1,
  A.init = NULL,
  B.init = NULL,
  phi.init = NULL,
  maxit = 250,
  tol = 1e-05,
  learning_rate = NULL,
  learning_rate.args = NULL,
  verbose = T
)
}
\arguments{
\item{Y:}{a matrix of dimensions n x q}

\item{q:}{the number of latent variables}

\item{X:}{either 0 (no covariates, no intercept), 1 (no covariates but an intercept), or a matrix of covariates. If the latter and an intercept is desired, it must be included in X as a column of 1.}

\item{family:}{one of "SA" or "SP"}

\item{H:}{how many samples of Z to draw: if in (0, 1), then a batch method is used and only that proportion of data is used to estimate the model.}
}
\description{
Returns a fitted object of class "fastgllvm".
}
