% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_binom}
\alias{gen_binom}
\title{Generate multivariate binomial data conditional on the linear parameter}
\usage{
gen_binom(linpar, size, nobs, p)
}
\description{
Generate multivariate binomial data conditional on the linear parameter
}
