% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_binom_fast}
\alias{gen_binom_fast}
\title{Generate multivariate binomial data conditional on the linear parameter}
\usage{
gen_binom_fast(linpar, size, nobs, p)
}
\description{
This is experimental and is faster than \code{gen_binom}.
}
