% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_fastgllvm}
\alias{gen_fastgllvm}
\title{Generates data from a GLLVM model.}
\usage{
gen_fastgllvm(
  n = 100,
  p = 5,
  q = 1,
  k = 1,
  family = "gaussian",
  A = NULL,
  B = NULL,
  phi = NULL,
  Z = NULL,
  X = NULL,
  intercept = T
)
}
\arguments{
\item{n}{the number of observations to draw}

\item{p}{the number of manifest variables}

\item{q}{the number of latent variables}

\item{k}{the number of covariates common to all responses, k=0 yields an intercept of 0, while k=1 yields a random intercept for each response, and k>1 generates covariates.}

\item{family}{one of "normal", "bernoulli", or a vector of both of size p, specifying individual responses' distribution}

\item{par}{optional parameters in a list}

\item{sigma}{sigma}
}
\value{
a list corresponding to the model
}
\description{
Returns a gllvm model of class fastglllvm with simulated data..
}
