% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gllvm.R
\name{gen_gllvm}
\alias{gen_gllvm}
\title{Generates data from a GLLVM model.}
\usage{
gen_gllvm(n, mu, sigma)
}
\arguments{
\item{n}{the number of observations to draw}

\item{mu}{mean}

\item{sigma}{sigma}
}
\value{
gllvm a list
}
\description{
Nothing to say
}
