% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gllvmprime.R
\name{gen_gllvmprime}
\alias{gen_gllvmprime}
\title{Generates a GLLVM model.}
\usage{
gen_gllvmprime(
  nobs = 100,
  p = 5,
  q = 1,
  k = 1,
  family = "gaussian",
  A = NULL,
  B = NULL,
  phi = NULL,
  X = NULL,
  Z = NULL,
  intercept = T,
  scale = 1,
  miss.prob = 0
)
}
\arguments{
\item{nobs}{the number of observations to draw}

\item{p}{the number of manifest variables}

\item{q}{the number of latent variables}

\item{k}{the number of covariates common to all responses, k=0 yields an intercept of 0, while k=1 yields a random intercept for each response, and k>1 generates covariates.}

\item{family}{one of "normal", "bernoulli", or a vector of both of size p, specifying individual responses' distribution}

\item{par}{optional parameters in a list}

\item{sigma}{sigma}
}
\value{
a list corresponding to the model
}
\description{
Returns an (unfitted) gllvm model of class "fastglllvm" with simulated data.
Because it is a gllvmprime object, it has components such as "hist" that are NULL
since no fit actually took place.
}
\details{
This function is useful for experimentation.
}
