% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_norm}
\alias{gen_norm}
\title{Generate multivariate Normal data conditional on the linear parameter}
\usage{
gen_norm(linpar, phi, nobs, p)
}
\description{
Generate multivariate Normal data conditional on the linear parameter
}
