% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gen_par}
\alias{gen_par}
\title{Generates parameters for a gllvm}
\usage{
gen_par(
  p,
  q,
  k = 0,
  family = "normal",
  A = NULL,
  B = NULL,
  Psi = NULL,
  scale = 0.5
)
}
\description{
returns a list of parameters.
}
