% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_poisson}
\alias{gen_poisson}
\title{Generate multivariate Poisson data conditional on the linear parameter}
\usage{
gen_poisson(linpar, nobs, p)
}
\description{
Generate multivariate Poisson data conditional on the linear parameter
}
