% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgllvm.R
\name{generate_families}
\alias{generate_families}
\title{Generate the families list as used in the other functions}
\usage{
generate_families(family, p)
}
\description{
Generate the families list as used in the other functions
}
