% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_y}
\alias{generate_y}
\title{Generates responses from a gllvm, given all parameters and values}
\usage{
generate_y(linpar, phi, families, nobs, p, q)
}
\arguments{
\item{A:}{loadings}

\item{B:}{beta}

\item{phi:}{scale parameters}

\item{Z:}{nobs x q matrix of latent variables}

\item{X:}{nobs x k design matrix}

\item{families:}{a named list with names, c("gaussian", "binomial", "poisson", "negbin"), of indices corresponding to the family of the responses.}

\item{known_par:}{a named list with names, c("gaussian", "binomial", "poisson", "negbin"), of known parameters (used for the size "n" of binomial responses), appearing in the same order as the corresponding element in \code{families}}
}
\value{
a list corresponding to the model
}
\description{
Returns a matrix Y
}
