% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_Z_functionfactory}
\alias{generate_Z_functionfactory}
\title{Returns a function used to generate Z. This is solely used within gllvm.}
\usage{
generate_Z_functionfactory(nobs, q, H = 1, method = "SA")
}
\arguments{
\item{method:}{one of "SA" (always random), "SP" (fixed)}

\item{H:}{number of draws to sample, usually set to 1 for SA and more for SP}
}
\value{
a function that returns a list of generated Z.
}
\description{
Returns a function used to generate Z. This is solely used within gllvm.
}
