% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{get_expectations}
\alias{get_expectations}
\title{Returns the (Monte-Carlo) approximations of E\link{Y} and E\link{Y Y}K.}
\usage{
get_expectations(A, B, X, Z, K, phi, family)
}
\arguments{
\item{A:}{the matrix of loadings}

\item{B:}{the matrix of fixed effect coefficients}

\item{X:}{the matrix of covariates}

\item{Z:}{a list of H generated Z samples}

\item{K:}{the matrix so that Z = K Y}

\item{family:}{an object of class "family"}
}
\description{
Returns the (Monte-Carlo) approximations of E\link{Y} and E\link{Y Y}K.
}
