% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_Psi}
\alias{get_Psi}
\title{Returns the psi function approximated by H monte carlo samples}
\usage{
get_Psi(Y, Y.c, A, B, phi, X, family, generate_Z)
}
\arguments{
\item{generate_Z}{a function returned by generate_Z_functionfactory that returns a list}

\item{Y:}{uncentered data}

\item{Y.c:}{Y that are CENTERED, they must have an attribute named \code{scaled:center} that represent the removed mean}

\item{A:}{the matrix of loadings}

\item{B:}{the matrix of fixed effect coefficients}

\item{phi:}{the vector of response-specific scale parameters}

\item{X:}{the matrix of covariates}

\item{family:}{an object of class "family"}
}
\description{
Returns the psi function approximated by H monte carlo samples
}
