% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gllvmprime.fit.R
\name{gllvmprime.fit}
\alias{gllvmprime.fit}
\title{Title}
\usage{
gllvmprime.fit(fg, parameters.init = NULL, controls)
}
\arguments{
\item{fg:}{a gllvmprime model}

\item{parameters.init:}{a list of initial values for the parameters}

\item{controls:}{a list of controls for the fitting process}
}
\value{
a gllvmprime object fitted to the data
}
\description{
Title
}
