% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gllvmprime.R
\name{gllvmprime}
\alias{gllvmprime}
\title{Fit a GLLVM using an M-estimator a stochatic approximation algorithm.}
\usage{
gllvmprime(
  Y,
  q = 1,
  family = "gaussian",
  X = NULL,
  intercept = T,
  Z.init = NULL,
  parameters.init = NULL,
  method = "full",
  verbose = F,
  hist = 100,
  hessian = T,
  use_signs = F,
  trim = 0.2,
  alpha = 0.2,
  batch_size = nrow(Y),
  maxit = 100,
  H = 1,
  H.seed = NULL,
  rescale = T
)
}
\arguments{
\item{Y:}{a \code{n} times \code{p} matrix, organized with the observations in the rows and the responses in the columns}

\item{q:}{the number of factors}

\item{family:}{specifies the families. See below for more information.}

\item{X:}{either \code{NULL} or a \code{n} timeis \code{k} matrix of covariates.}

\item{intercept:}{a boolean (default:TRUE) indicating whether an intercept should be included in the model. If \code{X} is supplied and an intercept is desired, the first column of \code{X} must be a vector of ones to model the intercept.}

\item{gradient_function:}{one of: "simple", "full", or a user-defined function.}

\item{parameters.init:}{a list containing "A", "B", and "phi".}

\item{hessian:}{one of "median", "F", or "T"}

\item{hist:}{the last \code{hist} iterations are kept for the graph. Do not keep any if set to NULL.}
}
\description{
Fits a gllvm model.
}
\details{
The implemented families currently are "gaussian", "binomial", or "poisson.
The \code{family} argument can be one of the following:
\itemize{
\item a string, the name of one of the implemented families.
\item a vector of strings of size \code{p}, each of which being the name of the implemented families.
\item a list: \code{list(objects, id, vec)}, where \code{objects} is a named list of objects of type "family",
\code{id} is a named list of integer vectors specifying the column numbers corresponding, whose names correspond to those of \code{objects}, and
\code{vec} is a named list of string vectors whose elements are the names corresponding to the \code{objects}$family name. This can be useful to specify
different link functions for the families.
}
}
\examples{
# Setting: declare the types of the response (in the order in which they appear)
family=c(rep("poisson", 50), rep("gaussian", 50), rep("binomial", 50))
q <- 2
p <- length(family)
set.seed(1234)
# Simulate data
data <- gen_gllvmprime(nobs=500, p=p, q=q, family=family, k=1, intercept=1, miss.prob = 0)
# Fit the data
fit <- gllvmprime(data$Y, q = q, X=data$X, family=family,  intercept = 1, controls = list(minit=20, maxit=100,alpha=1, eps=1e-3))
# Evaluate the fit
plot(fit)
# If necessary, continue the fitting process:
fit <- update(fit)
# Evaluate the fit
plot(fit)
}
