% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intialize.R
\name{initialize_learning_rate}
\alias{initialize_learning_rate}
\title{Function factory to return a function that generates a learning rate as a function of i}
\usage{
initialize_learning_rate(maxit, learning_rate.args = list(method = "spall"))
}
\arguments{
\item{method:}{the name of the method to use. One of "constant", "lin", "exp", "linexp", "spall".}

\item{method.args:}{list of method arguments}
}
\description{
Function factory to return a function that generates a learning rate as a function of i
}
