% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intialize.R
\name{initialize_parameters_delete}
\alias{initialize_parameters_delete}
\title{Given a fastgllvm object, compute good intial values.}
\usage{
initialize_parameters_delete(fastgllvm, target = NULL, rescale = F)
}
\arguments{
\item{fastgllvm:}{an object of class ''fastgllvm''}

\item{target:}{if non NULL, must be a matrix of loadings: will perform a Procrustes rotation to target}

\item{rescale:}{if true, both Z and A will be rescaled so that Z\%*\%t(A) remains unchanged but Z has variance identity.}
}
\description{
Given a fastgllvm object, compute good intial values.
}
