% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{MD}
\alias{MD}
\title{Compute the Mean Deviance of a fitted model}
\usage{
MD(Y, mu, family)
}
\arguments{
\item{mu:}{matrix of estimated means}
}
\description{
Compute the Mean Deviance of a fitted model
}
