% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{MPE}
\alias{MPE}
\title{Compute the mean procrustes error between two loading matrices.}
\usage{
MPE(A1, A2, rotate = T)
}
\arguments{
\item{A1:}{matrix of estimated loadings}

\item{A2:}{true matrix of loadings}
}
\description{
Compute the mean procrustes error between two loading matrices.
}
