% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{MSE}
\alias{MSE}
\title{Compute the mean squared error}
\usage{
MSE(B1, B2)
}
\arguments{
\item{B1:}{matrix of estimated coefficients}

\item{B2:}{true matrix of coefficients}
}
\description{
Compute the mean squared error
}
