% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{new_fastgllvm}
\alias{new_fastgllvm}
\title{Generates a fastgllvm object}
\usage{
new_fastgllvm(A, B, phi, family, X)
}
\arguments{
\item{A}{the matrix of loadings.}

\item{B}{the matrix of fixed effect coefficients, of dimensions p * k}

\item{phi}{a vector of scale parameters.}

\item{X}{either 0 (no covariates, no intercept), 1 (an intercept), or a matrix of n * k covariates (with, possibly, the first column of 1s being an intercept)}
}
\value{
a list corresponding to the model
}
\description{
Generates a fastgllvm object
}
