% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{predict.fastgllvm}
\alias{predict.fastgllvm}
\title{Predict the latent variable Z from a gllvm, assuming it is multivariate normally distributed.}
\usage{
\method{predict}{fastgllvm}(fastgllvm)
}
\arguments{
\item{fastgllvm}{an object of class "fastgllvm"}
}
\value{
a n x q matrix of factors
}
\description{
Predict the latent variable Z from a gllvm, assuming it is multivariate normally distributed.
}
