% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{simulate.gllvmprime}
\alias{simulate.gllvmprime}
\title{Simulate new data from a given gllvmprime object}
\usage{
\method{simulate}{gllvmprime}(gllvmprime, nsim = 1, conditional = F, return_object = F)
}
\arguments{
\item{index:}{numeric vector indicating which observation unit to simulate, useful when covariate information X is provided.}

\item{value:}{a gllvmprime object with the simulations}

\item{conditional:}{should the simulation be conditional on gllvmprime$Z?}
}
\description{
Simulate new data from a given gllvmprime object
}
