% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{subset.fastgllvm}
\alias{subset.fastgllvm}
\title{Subsets the fastgllvm object to only include observational units indicated in \code{index}}
\usage{
\method{subset}{fastgllvm}(fastgllvm, index)
}
\arguments{
\item{index:}{numeric vector indicating which observations to keep}
}
\description{
Subsets the fastgllvm object to only include observational units indicated in \code{index}
}
