% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{subset.gllvmprime}
\alias{subset.gllvmprime}
\title{Subsets the gllvmprime object to only include observational units indicated in \code{index}}
\usage{
\method{subset}{gllvmprime}(gllvmprime, index)
}
\arguments{
\item{index:}{numeric vector indicating which observations to keep}
}
\description{
Subsets the gllvmprime object to only include observational units indicated in \code{index}
}
