% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{update.fastgllvm}
\alias{update.fastgllvm}
\title{Update the fit with another round of stochastic approximation}
\usage{
\method{update}{fastgllvm}(fg, ...)
}
\arguments{
\item{...:}{named arguments, same as in the call to fastgllvm}
}
\description{
Update the fit with another round of stochastic approximation
}
