% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{update.gllvmprime}
\alias{update.gllvmprime}
\title{Update the fit with another round of stochastic approximation}
\usage{
\method{update}{gllvmprime}(fg, ...)
}
\arguments{
\item{...:}{named arguments, same as in the call to gllvmprime}
}
\description{
Update the fit with another round of stochastic approximation
}
