% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgllvm.R
\name{validate_fastgllvm}
\alias{validate_fastgllvm}
\title{Check whether the parameters are correctly specified}
\usage{
validate_fastgllvm(fastgllvm)
}
\description{
Check whether the parameters are correctly specified
}
