% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gllvmprime.R
\name{validate_gllvmprime}
\alias{validate_gllvmprime}
\title{Validates a gllvmprime object}
\usage{
validate_gllvmprime(gllvmprime)
}
\arguments{
\item{gllvmprime:}{a gllvmprime object

Strives to find errors in the gllvmprime object to simplify debugging.
Check whether the parameters are correctly specified}
}
\description{
Validates a gllvmprime object
}
