% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgllvm.fit.R
\name{fastgllvm.fit}
\alias{fastgllvm.fit}
\title{Fit a fastgllvm object for fitting a gllvm model.}
\usage{
fastgllvm.fit(fg, parameters.init = NULL, controls)
}
\description{
Fit a fastgllvm object for fitting a gllvm model.
}
