% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgllvm.R
\name{fastgllvm.fit}
\alias{fastgllvm.fit}
\title{Fit a fastgllvm object for fitting a gllvm model.}
\usage{
fastgllvm.fit(
  fg,
  method,
  parameters.init = NULL,
  controls = NULL,
  verbose = F,
  hist = F,
  median = F
)
}
\description{
Fit a fastgllvm object for fitting a gllvm model.
}
