% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfactoranalysis.R
\name{ffa}
\alias{ffa}
\title{Computes the maximum likelihood estimator for factor analysis}
\usage{
ffa(
  Y,
  q,
  maxiter = 100,
  eps = 1e-04,
  savepath = F,
  verbose = T,
  iteratively_update_Psi = T,
  rotate_updates = F
)
}
\description{
Computes the maximum likelihood estimator for factor analysis
}
