% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{predict.fastgllvm}
\alias{predict.fastgllvm}
\title{Predict the latent variable Z from a gllvm, assuming it is multivariate normally distributed.}
\usage{
\method{predict}{fastgllvm}(f, method = c("fastgllvm", "glm", "glmnet"))
}
\arguments{
\item{f:}{an object of class "fastgllvm"}

\item{method:}{one of "glm" or "ridge" (default)}
}
\value{
a n x q matrix of factors
}
\description{
Predict the latent variable Z from a gllvm, assuming it is multivariate normally distributed.
}
